requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5

# Web scraping
requests>=2.30.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0

# Text processing and chunking
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
tiktoken>=0.5.0
nltk>=3.8.0

# Vector databases and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0

# Data processing
pandas>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Frontend
streamlit>=1.25.0

# Evaluation and metrics
scikit-learn>=1.3.0
rouge-score>=0.1.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
pydantic-settings>=2.0.0

# Google Gemini
google-generativeai>=0.3.0

# Additional utilities
httpx>=0.24.0
lxml>=4.9.0
html5lib>=1.1
